pyproject.toml
[tool.poetry]
name = "ai-chatbot-service"
version = "1.0.0"
description = "A sophisticated AI chatbot service with real-time capabilities"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.4.2"}
pydantic-settings = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
redis = {extras = ["hiredis"], version = "^5.0.1"}
tenacity = "^8.2.3"
prometheus-client = "^0.19.0"
python-json-logger = "^2.0.7"
psutil = "^5.9.6"
websockets = "^12.0"
argon2-cffi = "^23.1.0"
alembic = "^1.12.1"
httpx = "^0.25.1"
PyJWT = "^2.8.0"
APScheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"
flake8 = "^6.1.0"
faker = "^19.13.0"
aioresponses = "^0.7.5"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app --cov-report=term-missing" 